+++ database/database.go
@@ -104,6 +104,23 @@ type Database interface {
 	FormatTime(t time.Time) string
 	ParseTime(t string) (time.Time, error)
 	DbType() string
+	GormDB() *gorm.DB
+	ChunkSize() int
+}
+
+func (it *Db) ChunkSize() int {
+	switch it.Database.Dialect().GetName() {
+	case "mysql":
+		return 3000
+	case "postgres":
+		return 3000
+	default:
+		return 100
+	}
+}
+
+func (it *Db) GormDB() *gorm.DB {
+	return it.Database
 }
 
 func (it *Db) DbType() string {