+++ frontend/src/languages/index.js
@@ -1,0 +1,15 @@
+import english from "./english"
+import spanish from "./spanish"
+import german from "./german"
+import russian from "./russian";
+import french from "./french";
+
+const language = {
+  en: english,
+  es: spanish,
+  de: german,
+  ru: russian,
+  fr: french,
+}
+
+export default language