+++ notifiers/discord.go
@@ -39,7 +39,7 @@ var Discorder = &discord{&notifications.Notification{
 
 // Send will send a HTTP Post to the discord API. It accepts type: []byte
 func (d *discord) sendRequest(msg string) error {
-	_, _, err := utils.HttpRequest(Discorder.GetValue("host"), "POST", "application/json", nil, strings.NewReader(msg), time.Duration(10*time.Second), true)
+	_, _, err := utils.HttpRequest(Discorder.GetValue("host"), "POST", "application/json", nil, strings.NewReader(msg), time.Duration(10*time.Second), true, nil)
 	return err
 }
 
@@ -63,7 +63,7 @@ func (d *discord) OnSuccess(s *services.Service) error {
 func (d *discord) OnTest() (string, error) {
 	outError := errors.New("Incorrect discord URL, please confirm URL is correct")
 	message := `{"content": "Testing the discord notifier"}`
-	contents, _, err := utils.HttpRequest(Discorder.Host, "POST", "application/json", nil, bytes.NewBuffer([]byte(message)), time.Duration(10*time.Second), true)
+	contents, _, err := utils.HttpRequest(Discorder.Host, "POST", "application/json", nil, bytes.NewBuffer([]byte(message)), time.Duration(10*time.Second), true, nil)
 	if string(contents) == "" {
 		return "", nil
 	}