+++ types/failures/samples.go
@@ -4,19 +4,19 @@ import (
 	"fmt"
 	"github.com/statping/statping/types"
 	"github.com/statping/statping/utils"
+	gormbulk "github.com/t-tiger/gorm-bulk-insert/v2"
 	"time"
 )
 
 var (
-	log = utils.Log
+	log = utils.Log.WithField("type", "failure")
 )
 
 func Samples() error {
+	log.Infoln("Inserting Sample Service Failures...")
 	createdAt := utils.Now().Add(-3 * types.Day)
 
 	for i := int64(1); i <= 4; i++ {
-		tx := db.Begin()
-
 		f1 := &Failure{
 			Service:   i,
 			Issue:     "Server failure",
@@ -33,21 +33,20 @@ func Samples() error {
 
 		log.Infoln(fmt.Sprintf("Adding %v Failure records to service", 400))
 
+		var records []interface{}
 		for fi := 0.; fi <= float64(400); fi++ {
 			failure := &Failure{
 				Service:   i,
 				Issue:     "testing right here",
 				CreatedAt: createdAt.UTC(),
 			}
-
-			tx = tx.Create(&failure)
+			records = append(records, failure)
 			createdAt = createdAt.Add(35 * time.Minute)
 		}
-		if err := tx.Commit().Error(); err != nil {
+		if err := gormbulk.BulkInsert(db.GormDB(), records, db.ChunkSize()); err != nil {
 			log.Error(err)
 			return err
 		}
 	}
-
 	return nil
 }