+++ utils/utils_windows.go
@@ -2,7 +2,10 @@ package utils
 
 import (
 	"errors"
+	"fmt"
 	"os"
+	"os/exec"
+	"strings"
 )
 
 func DirWritable(path string) (bool, error) {
@@ -20,4 +23,19 @@ func DirWritable(path string) (bool, error) {
 	}
 
 	return true, nil
+}
+
+func Ping(address string, secondsTimeout int) error {
+	ping, err := exec.LookPath("ping")
+	if err != nil {
+		return err
+	}
+	out, _, err := Command(ping, address, "-n 1", fmt.Sprintf("-w %v", secondsTimeout*1000))
+	if err != nil {
+		return err
+	}
+	if strings.Contains(out, "Destination Host Unreachable") {
+		return errors.New("destination host unreachable")
+	}
+	return nil
 }